Update version 1.1.9: Subject Identifier too long / Hidden hover panel

+++ R/updates_variables_selection_csvA.R
@@ -32,7 +32,7 @@ updates_variables_selection_csvA <- function(
 
           A2 <- numeric_to_integer(A)
 
-          integers_A <- names(which(unlist(lapply(A2,is.integer))))
+          integers_A <- names(which(unlist(lapply(A2,is.numeric))))
 
           if (shiny::isRunning()) {
             shiny::updateSelectInput(
