+++ R/updates_variables_selection_csvB.R
@@ -32,7 +32,7 @@ updates_variables_selection_csvB <- function(
 
           B2 <- numeric_to_integer(B)
 
-          integers_B <- names(which(unlist(lapply(B2,is.integer))))
+          integers_B <- names(which(unlist(lapply(B2,is.numeric))))
 
           if (shiny::isRunning()) {
             shiny::updateSelectInput(
